+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Add Camera preview library which shows camera preview in Android.
+- Add Camera preview library which shows camera preview in iOS.
 
 ## [0.3.0] - 2025-04-22